Started the refactor of the Betrayal UI
