+++ src/Sidekick.Localization/Leagues/Betrayal/BetrayalResources.de.resx
@@ -333,7 +333,7 @@ Skarabäus</value>
     <value>Perandus
 Münzen</value>
   </data>
-  <data name="JanusTransportaion" xml:space="preserve">
+  <data name="JanusTransportation" xml:space="preserve">
     <value>Qualitäts
 Währung</value>
   </data>
