+++ src/Sidekick.Localization/Leagues/Betrayal/BetrayalResources.zh-tw.resx
@@ -229,7 +229,7 @@
     <value>裂痕
 聖甲蟲</value>
   </data>
-  <data name="JanusTransportaion" xml:space="preserve">
+  <data name="JanusTransportation" xml:space="preserve">
     <value>品質通貨</value>
   </data>
   <data name="JanusFortification" xml:space="preserve">
