+++ src/Sidekick/Windows/LeagueOverlay/Betrayal/Agent.xaml
@@ -1,9 +1,9 @@
 <UserControl
-    x:Class="Sidekick.Windows.LeagueOverlay.Betrayal.Controls.Agent"
+    x:Class="Sidekick.Windows.LeagueOverlay.Betrayal.Agent"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="clr-namespace:Sidekick.Windows.LeagueOverlay.Betrayal.Controls"
+    xmlns:local="clr-namespace:Sidekick.Windows.LeagueOverlay.Betrayal"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     d:DesignHeight="450"
     d:DesignWidth="800"
