+++ src/Sidekick/Windows/LeagueOverlay/Betrayal/Agent.xaml.cs
@@ -6,17 +6,17 @@ using Sidekick.UI.Leagues.Betrayal;
 namespace Sidekick.Windows.LeagueOverlay.Betrayal
 {
     /// <summary>
-    /// Interaction logic for Betrayal.xaml
+    /// Interaction logic for Agent.xaml
     /// </summary>
     [DependencyProperty]
-    public partial class Betrayal : UserControl
+    public partial class Agent : UserControl
     {
-        public Betrayal()
+        public Agent()
         {
             InitializeComponent();
             DataContext = this;
         }
 
-        public BetrayalLeague League { get; set; }
+        public BetrayalAgent Item { get; set; }
     }
 }
