--- src/Sidekick/Windows/LeagueOverlay/Betrayal/Controls/Reward.xaml.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-
-namespace Sidekick.Windows.LeagueOverlay.Betrayal.Controls
-{
-    /// <summary>
-    /// Interaction logic for Reward.xaml
-    /// </summary>
-    public partial class Reward : UserControl
-    {
-        public Reward()
-        {
-            InitializeComponent();
-        }
-    }
-}
