+++ src/Sidekick/Windows/LeagueOverlay/Betrayal/Reward.xaml
@@ -1,5 +1,5 @@
 <UserControl
-    x:Class="Sidekick.Windows.LeagueOverlay.Betrayal.Betrayal"
+    x:Class="Sidekick.Windows.LeagueOverlay.Betrayal.Reward"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -8,13 +8,5 @@
     d:DesignHeight="450"
     d:DesignWidth="800"
     mc:Ignorable="d">
-    <WrapPanel>
-        <ItemsControl ItemsSource="{Binding Path=League.Agents}">
-            <ItemsControl.ItemTemplate>
-                <DataTemplate>
-                    <local:Agent Item="{Binding Path=.}" />
-                </DataTemplate>
-            </ItemsControl.ItemTemplate>
-        </ItemsControl>
-    </WrapPanel>
+    <Grid />
 </UserControl>
