+++ src/Sidekick/Windows/LeagueOverlay/Betrayal/Reward.xaml.cs
@@ -1,0 +1,15 @@
+using System.Windows.Controls;
+
+namespace Sidekick.Windows.LeagueOverlay.Betrayal
+{
+    /// <summary>
+    /// Interaction logic for Reward.xaml
+    /// </summary>
+    public partial class Reward : UserControl
+    {
+        public Reward()
+        {
+            InitializeComponent();
+        }
+    }
+}
